backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.155.0
celery[redis]==5.4.0
redis==5.2.1
itsdangerous==2.2.0
anthropic==0.40.0
python-multipart==0.0.20
